changed mixture amount, from integer to double.

+++ src/main/java/be/uantwerpen/labplanner/Model/Experiment.java
@@ -24,7 +24,7 @@ public class Experiment extends AbstractPersistable<Long> {
     @Column(name = "expMixtureComment")
     private String mixtureComment;
     @Column(name = "expmixtureAmount")
-    private int mixtureAmount;
+    private double mixtureAmount;
 
 
     @OneToMany
@@ -139,11 +139,11 @@ public class Experiment extends AbstractPersistable<Long> {
         return steps;
     }
 
-    public int getMixtureAmount() {
+    public double getMixtureAmount() {
         return mixtureAmount;
     }
 
-    public void setMixtureAmount(int mixtureAmount) {
+    public void setMixtureAmount(double mixtureAmount) {
         this.mixtureAmount = mixtureAmount;
     }
